refactor(ui reducers): make ui.ts file entrypoint for imports and handle platform specific logic internally
