+++ app/actions/platformSpecific/ui.electron.ts
@@ -1,5 +1,5 @@
 import { ipcRenderer } from 'electron'
-import { UI_SIGNOUT } from '../../reducers/ui.TARGET_PLATFORM'
+import { UI_SIGNOUT } from '../../reducers/ui'
 
 export const signout = () => {
   ipcRenderer.send('block-menus', true)
